requirements.txt
streamlit 
python-dotenv 
langchain-google-genai 
langchain-astradb 
pypdf
langchain-community